Hack to make sure the first-loaded Mallard file uses the Mallard schema
Since our function for nxml-mode-hook runs *after* nxml-mode is initialized for a buffer,
it means that by the time we add the mallard-locating-schema.xml to rng-schema-locating-files,
the process of auto-detecting an XML file's schema has already run, without our newly-added
schema file.

Subsequent visits of a Mallard file will work because by then rng-schema-locating-files is
already tweaked with our extra file.

But for now, we tweak the variable and call rng-auto-set-schema-and-validate explicitly.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>